Move InteractHandlers to its own module
