+++ rs/src/ui/pause_menu.rs
@@ -23,8 +23,8 @@ use engine::{
 		layout::{ExpandToFitChildren, LineUpChildren},
 		text::UiFonts,
 		widgets::{
-			borders::Border, dbg_event, focus_state_colors, focus_state_emissive, on_ok,
-			CuboidPanel, InteractHandlers, Node3d, Text3d, WidgetShape,
+			borders::Border,
+			CuboidPanel, Node3d, Text3d, WidgetShape,
 		},
 		Fade, FadeCommands, GlobalUi, MenuRef, MenuStack, UiMat, UiMatBuilder,
 	},
@@ -36,6 +36,7 @@ use rapier3d::geometry::SharedShape;
 use smallvec::smallvec;
 use std::ops::ControlFlow;
 use tiny_bail::prelude::r;
+use engine::ui::interact::{dbg_event, focus_state_colors, focus_state_emissive, on_ok, InteractHandlers};
 
 pub struct PauseMenuPlugin;
